Tidy types and add VersionedBindings type
