+++ src/Language/Mimsa/Types.hs
@@ -6,6 +6,7 @@
 
 module Language.Mimsa.Types
   ( module Language.Mimsa.Types.Name,
+    module Language.Mimsa.Types.ExprHash,
     module Language.Mimsa.Types.AST,
     module Language.Mimsa.Types.Store,
     module Language.Mimsa.Types.TypeError,
@@ -24,14 +25,20 @@ module Language.Mimsa.Types
     module Language.Mimsa.Types.Swaps,
     module Language.Mimsa.Types.UniVar,
     module Language.Mimsa.Types.Tui,
+    module Language.Mimsa.Types.Library,
+    module Language.Mimsa.Types.Bindings,
+    module Language.Mimsa.Types.StoreExpression,
   )
 where
 
 import Language.Mimsa.Types.AST
+import Language.Mimsa.Types.Bindings
 import Language.Mimsa.Types.Environment
 import Language.Mimsa.Types.Error
+import Language.Mimsa.Types.ExprHash
 import Language.Mimsa.Types.ForeignFunc
 import Language.Mimsa.Types.InterpreterError
+import Language.Mimsa.Types.Library
 import Language.Mimsa.Types.MonoType
 import Language.Mimsa.Types.Name
 import Language.Mimsa.Types.Printer
@@ -40,6 +47,7 @@ import Language.Mimsa.Types.ResolverError
 import Language.Mimsa.Types.Scheme
 import Language.Mimsa.Types.Scope
 import Language.Mimsa.Types.Store
+import Language.Mimsa.Types.StoreExpression
 import Language.Mimsa.Types.Substitutions
 import Language.Mimsa.Types.Swaps
 import Language.Mimsa.Types.Tui
