+++ src/Language/Mimsa/Types/ResolvedDeps.hs
@@ -7,7 +7,7 @@ import qualified Data.Map as M
 import qualified Data.Text as T
 import Language.Mimsa.Types.Name
 import Language.Mimsa.Types.Printer
-import Language.Mimsa.Types.Store
+import Language.Mimsa.Types.StoreExpression
 
 newtype ResolvedDeps = ResolvedDeps {getResolvedDeps :: Map Name StoreExpression}
