+++ src/Language/Mimsa/Types/ResolverError.hs
@@ -4,9 +4,9 @@ module Language.Mimsa.Types.ResolverError where
 
 import qualified Data.Map as M
 import qualified Data.Text as T
+import Language.Mimsa.Types.Bindings
 import Language.Mimsa.Types.Name
 import Language.Mimsa.Types.Printer
-import Language.Mimsa.Types.Store
 
 data ResolverError = MissingBinding Name Bindings
   deriving (Eq, Ord, Show)
